+++ graph/Graph.h
@@ -82,8 +82,8 @@ public:
 	void Export(const char *fname);
 	
 	int AddNode(node *);
-	node *GetNode(unsigned int num);
-	edge *GetEdge(unsigned int num);
+	node *GetNode(unsigned long num);
+	edge *GetEdge(unsigned long num);
 	void AddEdge(edge *);
 	edge *findDirectedEdge(unsigned int from, unsigned int to);
 	edge *FindEdge(unsigned int from, unsigned int to);
